Fix bug in ExponentialDecay where long was not cast to double.

+++ src/main/java/decayfunction/ExponentialDecay.java
@@ -23,7 +23,7 @@ public class ExponentialDecay extends DecayFunction {
      */
     @Override
     public double decay(double initial, long millisPassed) {
-        return initial * Math.pow(getDecayFactor(), millisPassed / 1000);
+        return initial * Math.pow(getDecayFactor(), (double) millisPassed / 1000);
     }
 
 }
